Feature: edit gif with various options and download it - 여러 옵션을 활용한 gif 편집 - 편집한 gif 다운로드
